requirements.txt
numpy
opencv-python
Pillow
torch
torchvision
tqdm
